requirements.txt
APScheduler==3.3.1
chardet==2.3.0
click==6.7
colorama==0.3.2
demjson==2.2.4
Flask==0.12
flask-mongoengine==0.9.2
Flask-SocketIO==2.7.0
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
funcsigs==1.0.2
futures==3.0.5
gi==1.2
html5lib==0.999
itsdangerous==0.24
Jinja2==2.9.5
legacy==0.1.6
MarkupSafe==1.0
mongoengine==0.11.0
MyApplication==0.1.0
ndg-httpsclient==0.3.2
pyasn1==0.1.7
pymongo==3.4.0
pyOpenSSL==0.13.1
pyserial==3.2.1
python-apt==0.9.3.12
pytz==2016.7
requests==2.12.1
RPi.GPIO==0.6.3
six==1.10.0
SQLAlchemy==1.1.4
tzlocal==1.3
urllib3==1.9.1
Werkzeug==0.12.1
WTForms==2.1
